feat: OpenCL GPU prefill offload, tool-call reliability, self-hosting updates
GPU prefill offload (Adreno, OpenCL):
- cl.c/cl.h: OpenCL engine ported from ma3's approach, using llama.cpp's
  production noshuffle Q4_K/Q6_K GEMM kernels adapted for qma's k-quants
  (i8mm-ordered Q4_K qs, LFM-ordered Q6_K, host-precomputed scale/min so
  the GPU does no scale arithmetic — the Adreno compiler miscompiles it)
- whole-vector tile output + GPU transpose kernel (component stores
  corrupt the accumulation on this driver); decode and streamed experts
  stay CPU; QMA_NO_CL=1 falls back; tests/cltest.c is the numerical gate
  (GPU matches CPU within ~0.1%% fp16 noise on all shapes)

Tool-call reliability:
- no grammar mask, one complete call per turn with forced <|im_end|>
- lenient <tool_call> parser: accepts dropped openers, inline values,
  improvised close tags (</KEY>, </arg...>) without breaking </dev/null
- malformed-call feedback now shows the exact format; temp default 0.8

Agent:
- enter() now physically moves the running binary to the new cwd (rename
  works on a running executable; /proc/self/exe follows), so the next
  self-hosted generation is written into the project the agent moved to;
  skipped inside /internal and across filesystems (EXDEV reported)

Also carries earlier uncommitted engine work: GGUF 4K-alignment repack,
selfctx session embedding, thermal drain, embedded internal tree.

+++ src/ecache.c
@@ -36,9 +36,13 @@ void qma_ecache_prof_delta(const char *tag){
     extern qma_ecache *g_prof_cache;
     c = g_prof_cache;
     fprintf(stderr, "[ecprof:%s] sync=%llu(%.1fms) wait=%llu(%.1fms) hit=%llu", tag,
-            g_ec_sync_cnt-p_s, g_ec_sync_ms-p_sm, g_ec_wait_cnt-p_w, g_ec_wait_ms-p_wm, g_ec_hit_cnt-p_h);
+            (unsigned long long)(g_ec_sync_cnt-p_s), g_ec_sync_ms-p_sm,
+            (unsigned long long)(g_ec_wait_cnt-p_w), g_ec_wait_ms-p_wm,
+            (unsigned long long)(g_ec_hit_cnt-p_h));
     if (c) fprintf(stderr, " ev=%llu spec=%llu miss=%llu",
-                   c->evictions-p_ev, c->spec_issued-p_sp, c->misses);
+                   (unsigned long long)(c->evictions-p_ev),
+                   (unsigned long long)(c->spec_issued-p_sp),
+                   (unsigned long long)(c->misses));
     fprintf(stderr, "\n");
     p_s=g_ec_sync_cnt; p_sm=g_ec_sync_ms; p_w=g_ec_wait_cnt; p_wm=g_ec_wait_ms; p_h=g_ec_hit_cnt;
     if (c) { p_ev=c->evictions; p_sp=c->spec_issued; }
